fix(dbus): build bus address when not detected
When a DBUS session is already open for the user (eg the current
user's session) but could not be detected from its process name,
directly build its dbus session bus address from /run/user.

Otherwise, a new dbus session would be created, but it could not
completely update the settings.